Throw a custom Exception on non existing value.
`factories` are not allowed to return null after migrating to null
safety. The recommended solution is to throw an exception instead, or
create a static method. In this case a static method is not possible
because static methods can not access type parameters of the class.
Furthermore, throwing an exception is probably the best solution anyway,
since, like the documentation says, a non existing value is an invalid
input.

Reference:
https://dart.dev/null-safety/faq#how-should-i-migrate-a-factory-that-can-return-null